pyproject.toml
[project]
name = "mit-web-development-final-year-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "swanpyaesone26",email = "joyboyop69@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12, <4.0"
dependencies = [
    "django (>=6.0.2,<7.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-filter (>=25.2,<26.0)",
    "django-unfold (>=0.81.0,<0.82.0)",
    "django-crispy-forms (>=2.5,<3.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "gunicorn (>=22.0.0,<23.0.0)",
    "boto3 (>=1.34.0,<2.0.0)",
    "django-storages (>=1.14.0,<2.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
